refactor(node): reorganize _next_tick.ts module
