+++ node/_next_tick.ts
@@ -1,147 +1,47 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
-
-/** https://nodejs.org/api/process.html#process_process_nexttick_callback_args */
+// Copyright Joyent, Inc. and other Node contributors.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a
+// copy of this software and associated documentation files (the
+// "Software"), to deal in the Software without restriction, including
+// without limitation the rights to use, copy, modify, merge, publish,
+// distribute, sublicense, and/or sell copies of the Software, and to permit
+// persons to whom the Software is furnished to do so, subject to the
+// following conditions:
+//
+// The above copyright notice and this permission notice shall be included
+// in all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
+// OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN
+// NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+// DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+// OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
+// USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 import { validateCallback } from "./_validators.ts";
-// FIXME(bartlomieju):
-// import { process } from "./process.ts";
+import { process } from "./process.ts";
+import { FixedQueue } from "./_fixed_queue.ts";
 
 // deno-lint-ignore no-explicit-any
 const core = ((Deno as any).core as any);
-
-// Currently optimal queue size, tested on V8 6.0 - 6.6. Must be power of two.
-const kSize = 2048;
-const kMask = kSize - 1;
-
-// The FixedQueue is implemented as a singly-linked list of fixed-size
-// circular buffers. It looks something like this:
-//
-//  head                                                       tail
-//    |                                                          |
-//    v                                                          v
-// +-----------+ <-----\       +-----------+ <------\         +-----------+
-// |  [null]   |        \----- |   next    |         \------- |   next    |
-// +-----------+               +-----------+                  +-----------+
-// |   item    | <-- bottom    |   item    | <-- bottom       |  [empty]  |
-// |   item    |               |   item    |                  |  [empty]  |
-// |   item    |               |   item    |                  |  [empty]  |
-// |   item    |               |   item    |                  |  [empty]  |
-// |   item    |               |   item    |       bottom --> |   item    |
-// |   item    |               |   item    |                  |   item    |
-// |    ...    |               |    ...    |                  |    ...    |
-// |   item    |               |   item    |                  |   item    |
-// |   item    |               |   item    |                  |   item    |
-// |  [empty]  | <-- top       |   item    |                  |   item    |
-// |  [empty]  |               |   item    |                  |   item    |
-// |  [empty]  |               |  [empty]  | <-- top  top --> |  [empty]  |
-// +-----------+               +-----------+                  +-----------+
-//
-// Or, if there is only one circular buffer, it looks something
-// like either of these:
-//
-//  head   tail                                 head   tail
-//    |     |                                     |     |
-//    v     v                                     v     v
-// +-----------+                               +-----------+
-// |  [null]   |                               |  [null]   |
-// +-----------+                               +-----------+
-// |  [empty]  |                               |   item    |
-// |  [empty]  |                               |   item    |
-// |   item    | <-- bottom            top --> |  [empty]  |
-// |   item    |                               |  [empty]  |
-// |  [empty]  | <-- top            bottom --> |   item    |
-// |  [empty]  |                               |   item    |
-// +-----------+                               +-----------+
-//
-// Adding a value means moving `top` forward by one, removing means
-// moving `bottom` forward by one. After reaching the end, the queue
-// wraps around.
-//
-// When `top === bottom` the current queue is empty and when
-// `top + 1 === bottom` it's full. This wastes a single space of storage
-// but allows much quicker checks.
-
-class FixedCircularBuffer {
-  bottom: number;
-  top: number;
-  list: undefined | Array<unknown>;
-  next: FixedCircularBuffer | null;
-
-  constructor() {
-    this.bottom = 0;
-    this.top = 0;
-    this.list = new Array(kSize);
-    this.next = null;
-  }
-
-  isEmpty() {
-    return this.top === this.bottom;
-  }
-
-  isFull() {
-    return ((this.top + 1) & kMask) === this.bottom;
-  }
-
-  push(data: unknown) {
-    this.list![this.top] = data;
-    this.top = (this.top + 1) & kMask;
-  }
-
-  shift() {
-    const nextItem = this.list![this.bottom];
-    if (nextItem === undefined) {
-      return null;
-    }
-    this.list![this.bottom] = undefined;
-    this.bottom = (this.bottom + 1) & kMask;
-    return nextItem;
-  }
-}
 
 interface Tock {
   callback: (...args: Array<unknown>) => void;
   args: Array<unknown>;
 }
-class FixedQueue {
-  head: FixedCircularBuffer;
-  tail: FixedCircularBuffer;
-
-  constructor() {
-    this.head = this.tail = new FixedCircularBuffer();
-  }
-
-  isEmpty() {
-    return this.head.isEmpty();
-  }
-
-  push(data: unknown) {
-    if (this.head.isFull()) {
-      // Head is full: Creates a new queue, sets the old queue's `.next` to it,
-      // and sets it as the new main queue.
-      this.head = this.head.next = new FixedCircularBuffer();
-    }
-    this.head.push(data);
-  }
-
-  shift() {
-    const tail = this.tail;
-    const next = tail.shift();
-    if (tail.isEmpty() && tail.next !== null) {
-      // If there is another queue, it forms the new tail.
-      this.tail = tail.next;
-    }
-    return next;
-  }
-}
 
 const queue = new FixedQueue();
 
 function runNextTicks() {
-  // FIXME(bartlomieju):
+  // FIXME(bartlomieju): Deno currently doesn't support
+  // unhandled rejections
   // if (!hasTickScheduled() && !hasRejectionToWarn())
   //   runMicrotasks();
   // if (!hasTickScheduled() && !hasRejectionToWarn())
   //   return;
+
   if (!core.hasTickScheduled()) {
     core.runMicrotasks();
   }
@@ -158,7 +58,7 @@ function processTicksAndRejections() {
   do {
     // deno-lint-ignore no-cond-assign
     while (tock = queue.shift()) {
-      // FIXME(bartlomieju):
+      // FIXME(bartlomieju): Deno currently doesn't support async hooks
       // const asyncId = tock[async_id_symbol];
       // emitBefore(asyncId, tock[trigger_async_id_symbol], tock);
 
@@ -186,20 +86,21 @@ function processTicksAndRejections() {
           }
         }
       } finally {
-        // FIXME(bartlomieju):
+        // FIXME(bartlomieju): Deno currently doesn't support async hooks
         // if (destroyHooksExist())
         // emitDestroy(asyncId);
       }
 
-      // FIXME(bartlomieju):
+      // FIXME(bartlomieju): Deno currently doesn't support async hooks
       // emitAfter(asyncId);
     }
     core.runMicrotasks();
-    // FIXME(bartlomieju):
+    // FIXME(bartlomieju): Deno currently doesn't support unhandled rejections
     // } while (!queue.isEmpty() || processPromiseRejections());
   } while (!queue.isEmpty());
+
   core.setHasTickScheduled(false);
-  // FIXME(bartlomieju):
+  // FIXME(bartlomieju): Deno currently doesn't support async hooks
   // setHasRejectionToWarn(false);
 }
 
@@ -220,10 +121,9 @@ export function nextTick<T extends Array<unknown>>(
 ) {
   validateCallback(callback);
 
-  // FIXME(bartlomieju):
-  // if (process._exiting) {
-  //   return;
-  // }
+  if (process._exiting) {
+    return;
+  }
 
   // TODO(bartlomieju): seems superfluous if we don't depend on `arguments`
   let args_;
@@ -249,16 +149,21 @@ export function nextTick<T extends Array<unknown>>(
   if (queue.isEmpty()) {
     core.setHasTickScheduled(true);
   }
-  // FIXME(bartlomieju):
+  // FIXME(bartlomieju): Deno currently doesn't support async hooks
   // const asyncId = newAsyncId();
   // const triggerAsyncId = getDefaultTriggerAsyncId();
+
   const tickObject = {
+    // FIXME(bartlomieju): Deno currently doesn't support async hooks
     // [async_id_symbol]: asyncId,
     // [trigger_async_id_symbol]: triggerAsyncId,
     callback,
     args: args_,
   };
+
+  // FIXME(bartlomieju): Deno currently doesn't support async hooks
   // if (initHooksExist())
   //   emitInit(asyncId, 'TickObject', triggerAsyncId, tickObject);
+
   queue.push(tickObject);
 }
